Rework arrows to be defined via macro and implement the 'some' arrows
This is a pretty big update that involved a lot of fiddling to get it right.

First, a macro ~? was created to handle single-term reduction inside the
arrow macros and allowing a more general arrow macro.

Second, looking up the definition of the 'and' macro in the R7RS standard
helped immensely with making the 'some' macros behave properly.

At this point, all the macros had almost identical definitions, so I threw it
all in a macro-defining macro to reduce code duplication.

N.B. It's probably possible to eliminate ~? with a conditional like the one
near the end of define-arrow.