Generalize define-arrow; change name. Define new wands.
An extra parameter and condition was added to define-arrow to allow it to also
be used for defining wands. As a result, its name was changed to
define-threading-macro to reflect its generality.

This change corrects ~<> to have the proper behavior when no <> appears in a
particular form.

Additionally, the following wands were defined:
- ~<>>
- some~<>
- some~<>>